Development

+++ app/api/auctionlistings.py
@@ -1,6 +1,7 @@
 from flask import Blueprint, request
 from flask_login import login_required, current_user
 import json
+from datetime import datetime
 from app.models import AuctionListing, db
 from app.forms import AuctionListingForm
 from app.api import validation_errors_to_error_messages
@@ -23,18 +24,19 @@ def get_auction_listing(auctionlisting_id):
         return {"errors": "Auctionlisting not found."}, 404
 
     if request.method == "DELETE":
-        # if not single_listing.check_owner(owner_id):
-        #     return {"errors": "Forbidden."}, 403
-        # else:
-        db.session.delete(single_listing)
-        db.session.commit()
-        return {"Success": "Listing deleted."}, 202
+        if not single_listing.check_owner(single_listing.owner_id):
+            return {"errors": "Forbidden."}, 403
+        else:
+            db.session.delete(single_listing)
+            db.session.commit()
+            return {"Success": "Listing deleted."}, 202
 
     if request.method == "PUT":
         form = AuctionListingForm()
         form["csrf_token"].data = request.cookies["csrf_token"]
         if form.validate_on_submit():
             single_listing.auction_deadline = form.data["auction_deadline"]
+            single_listing.last_update = datetime.utcnow()
             db.session.commit()
             return single_listing.to_safe_dict(), 200
         else:
